src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.6.0
anyascii==0.3.2
appdirs==1.4.4
argcomplete==3.3.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
bcrypt==4.1.3
Beaker==1.12.1
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
black==24.4.2
blivet==3.10.0
blivet-gui==2.5.0
boto3==1.35.12
botocore==1.35.12
Brlapi==0.8.5
cached-property==1.5.2
cachetools==5.3.3
cairocffi==1.6.1
CairoSVG==2.7.1
cattrs==23.2.3
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
crispy-tailwind==1.0.3
cryptography==41.0.7
cson==0.8
cssbeautifier==1.15.1
cssselect==1.1.0
cssselect2==0.7.0
cupshelpers==1.0
daphne==4.1.2
dasbus==1.7
dataclasses-json==0.6.7
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==5.0.6
django-active-link==0.1.8
django-appconf==1.0.6
django-celery-results==2.5.1
django-compressor==4.5.1
django-cotton==0.9.37
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-decouple==2.1
django-extensions==3.2.3
django-fbv==0.5.1
django-filter==24.2
django-htmx==1.17.3
django-js-routes==0.2.0
django-modelcluster==6.3
django-ninja==1.3.0
django-pandas==0.6.7
django-permissionedforms==0.1
django-qr-code==4.0.1
django-rest-framework==0.1.0
django-stubs-ext==5.0.4
django-taggit==5.0.1
django-tasks==0.4.0
django-template-partials==24.2
django-treebeard==4.7.1
django-unicorn==0.61.0
django-vite==3.0.4
django-widget-tweaks==1.5.0
django_components==0.81
djangorestframework==3.15.2
djlint==1.34.1
dnf==4.21.1
docker==6.1.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
draftjs-exporter==5.0.0
EditorConfig==0.12.4
esbonio==0.16.4
et-xmlfile==1.1.0
executing==2.0.1
Faker==26.0.0
fedora-third-party==0.10
ffmpeg==1.4
ffmpeg-python==0.2.0
file-magic==0.4.0
filelock==3.13.1
filetype==1.2.0
fros==1.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
gpg==1.23.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
html-tag-names==0.1.2
html-void-elements==0.1.0
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.5
humanize==3.13.1
hyperframe==5.2.0
hyperlink==21.0.0
icecream==2.1.3
idna==2.10
imagesize==1.4.1
incremental==24.7.2
inertia-django==0.6.0
inkex==1.3.1
ipython==8.26.0
iso639==0.1.4
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsbeautifier==1.15.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
kombu==5.3.7
l18n==2021.3
laces==0.1.1
langchain==0.2.12
langchain-community==0.2.5
langchain-core==0.2.29
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.98
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.3
louis==3.28.0
lsprotocol==2023.0.1
lxml==5.1.0
Mako==1.2.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
mtranslate==1.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nftables==0.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
olefile==0.47
openpyxl==3.1.2
orjson==3.10.5
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
Paste==3.7.1
pathspec==0.12.1
pexpect==4.9.0
pid==2.2.3
pillow==10.2.0
pillow_heif==0.18.0
platformdirs==4.2.2
plotly==5.20.0
ply==3.11
productmd==1.38
progressbar2==3.53.2
prompt_toolkit==3.0.47
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.22
pycrypto==2.6.1
pycups==2.0.4
pycurl==7.45.2
pydantic==2.8.2
pydantic_core==2.20.1
pyenchant==3.2.2
pygit2==1.14.0
pygls==1.3.1
Pygments==2.18.0
PyGObject==3.48.2
pygount==1.8.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pykickstart==3.52
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyparted==3.13.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyserial==3.5
PySocks==1.7.1
pyspellchecker==0.8.1
python-augeas==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-distutils-extra==2.39
python-dotenv==1.0.1
python-meh==0.51
python-pam==2.0.2
python-slugify==8.0.4
python-utils==3.7.0
pytz==2024.1
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.2
redis==5.0.6
referencing==0.31.1
regex==2023.12.25
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rfc3986==1.5.0
rich==13.7.1
rjsmin==1.2.2
rpds-py==0.18.1
rpm==4.19.1.1
rpmautospec==0.7.2
rpmautospec-core==0.1.5
ruff==0.4.10
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
scour==0.38.2
segno==1.6.1
selinux @ file:///builddir/build/BUILD/libselinux-3.6/src
sentence-transformers==3.0.1
sepolicy @ file:///builddir/build/BUILD/selinux-3.6/python/sepolicy
service-identity==24.1.0
setools==4.5.1
setuptools==69.0.3
shortuuid==1.0.13
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
slippers==0.6.2
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sos==4.7.2
soupsieve==2.5
sourcetypes==0.0.4
speg==0.3
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.32
sqlite-vec==0.1.1
sqlite-vss==0.1.2
sqlparse==0.5.0
stack-data==0.6.3
stripe==8.8.0
sweetify==2.3.1
sympy==1.13.2
systemd-python==235
telepath==0.3.1
Tempita==0.5.2
tenacity==8.2.3
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
Twisted==24.7.0
txaio==23.1.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.21.1
wagtail==6.2.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==3.0.4
whitenoise==6.6.0
Willow==1.8.0
xkbregistry==0.3
yarl==1.9.4
zope.interface==7.0.1
